Created some new helpers
Created a DataObject class to easily load in data from json to an
Object for later use.
Created a Grid and GridNode to store data in a grid.
Created a TileGrid to replace the TileMap and TileSprite to replace Tile
created a gulp extension to create a json file with all of the image
assets to easily load with PIXIjs’ loader